Create README.md
An organization screens candidates with an aptitude test prior to hiring. Data from employee annual performance reviews were collected if the employee was hired by the organization for at least 6 months. Inferential statistics was used to determine whether the test was successful on predicting employee performance. Additionally, the results from the test and performance ratings were analyzed to determine if there was any systemic bias due to employee race or gender.